% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkIntFormat}
\alias{checkIntFormat}
\title{A way of checking if a string is a valid file name.}
\usage{
checkIntFormat(s)
}
\arguments{
\item{s}{An integer}
}
\value{
logical: TRUE or FALSE
}
\description{
A copy from R's grDevices:::checkIntFormat because it is not exported there
}
\author{
R Core Development Team
}

