% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawnorm.R
\name{drawnorm}
\alias{drawnorm}
\title{draw a normal distribution with beautiful illustrations}
\usage{
drawnorm(
  mu = 0,
  sigma = 1,
  xlab = "A Normally Distributed Variable",
  ylab = "Probability Density",
  main,
  ps = par("ps"),
  ...
)
}
\arguments{
\item{mu}{The mu parameter}

\item{sigma}{The sigma parameter}

\item{xlab}{Label for x axis}

\item{ylab}{Label for Y axis}

\item{main}{Title for plot. OK to ignore this, we'll make a nice one for you}

\item{ps}{pointsize of text}

\item{...}{arguments passed to par}
}
\description{
This was developed for the R Working Example collection
in my website, pj.freefaculty.org/R/WorkingExamples
}
\examples{
drawnorm(mu = 10, sigma = 20)
drawnorm(mu= 0, sigma = 1)
drawnorm(mu = 102, sigma = 313)
drawnorm(mu = 0, sigma = 1, main = "A Standard Normal Distribution, N(0,1)",
         xlab = "X", ylab = "Density", ps = 7)
drawnorm(mu = 0, sigma = 1, ylab = "Density", ps = 14) 
}
\author{
Paul Johnson <pauljohn@ku.edu>
}
