% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outreg.R
\name{formatVC}
\alias{formatVC}
\title{formatter for merMod objects copied from lme4}
\usage{
formatVC(varc, digits = max(3, getOption("digits") - 2), comp = "Std.Dev.",
  formatter = format, ...)
}
\arguments{
\item{varc}{variance estimates}

\item{digits}{digits desired}

\item{comp}{what do you want}

\item{formatter}{a format function}

\item{...}{other arguments}
}
\value{
formatted text
}
\description{
R packaging started to complain about usage of non-exported
functions from packages. lme4 team said they might export
this function at some time in future. Until then, I need to copy it.
}
\author{
Doug Bates, Martin Machler, Ben Bolker, Stephen Walker
}
