% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCorrelatedData.R
\name{makeVec}
\alias{makeVec}
\title{makeVec for checking or creating vectors}
\usage{
makeVec(x = NULL, d = NULL)
}
\arguments{
\item{x}{A single value or a vector}

\item{d}{An integer, the desired size of the vector}
}
\value{
A vector of length d
}
\description{
This is a convenience for handling function arguments. If x is a
single value, it makes a vector of length d in which all values
are equal to x. If x is a vector, check that its length is d.
}
\author{
Paul E. Johnson <pauljohn@ku.edu>
}
