% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictOMatic.R
\name{model.data}
\alias{model.data}
\title{Create a "raw" (UNTRANSFORMED) data frame equivalent to the input
data that would be required to fit the given model.}
\usage{
model.data(model, ...)
}
\arguments{
\item{model}{A fitted regression model in which the data argument
is specified. This function will fail if the model was not fit
with the data option.}

\item{...}{Arguments passed to implementing methods.}
}
\value{
A data frame
}
\description{
This is a generic method. Unlike model.frame and
model.matrix, this does not return transformed variables. It deals
with regression formulae that have functions like poly(x, d) in
them. It differentiates x from d in those expressions. And it also
manages log(x + 10). The default method works for standarad R
regression models like lm and glm.
}
\author{
Paul Johnson \email{pauljohn@ku.edu}
}
