% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{print.summarize}
\alias{print.summarize}
\title{print method for output from summarize}
\usage{
\method{print}{summarize}(x, digits, ...)
}
\arguments{
\item{x}{Object produced by summarize}

\item{digits}{Decimal values to display, defaults as 2.}

\item{...}{optional arguments for print function.}
}
\value{
x, unchanged
Prints objects created by summarize
}
\description{
Be aware that the unrounded numeric matrix
is available as an attribute of the returned object.
This method displays a rounded, character-formatted
display of the numeric varibles.
}
