% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctable.R
\name{print.summary.pctable}
\alias{print.summary.pctable}
\title{print method for summary.pctable objects}
\usage{
\method{print}{summary.pctable}(x, ...)
}
\arguments{
\item{x}{a summary.pctable object}

\item{...}{Other arguments to print method}
}
\value{
Nothing is returned
}
\description{
prints pctab objects. Needed only to deal properly with quotes
}
\author{
Paul Johnson <pauljohn@ku.edu>
}
