% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNULL.R
\name{removeNULL}
\alias{removeNULL}
\title{Remove NULL values variables from a list}
\usage{
removeNULL(aList)
}
\arguments{
\item{aList}{A list}
}
\value{
Same list with NULL's removed
}
\description{
Unlike vectors, lists can hold objects with
value NULL. This gets rid of them.
}
\details{
This version is NOT recursive

plyr::rbind.fill uses an experimental function
that I choose to avoid. This is the "safe" version.
}
\examples{
## Note it is non-recursive, NULL remains in e
x <- list(a = rnorm(5), b = NULL, c = rnorm(5), d = NULL,
     e = list(f = rnorm(2), g = NULL))
x
removeNULL(x)
}
\author{
Paul Johnson
}
