\docType{package}
\name{rockchalk-package}
\alias{rockchalk}
\alias{"rockchalk-package"}
\alias{rockchalk-package}
\title{Miscellaneous regression functions}
\description{
  \tabular{ll}{ Package: \tab rockchalk\cr Type: \tab
  Package\cr Version: \tab 1.0\cr Date: \tab 2011-08-29\cr
  License: \tab GPL >= 3\cr LazyLoad: \tab yes\cr } ~~ An
  overview of how to use the package, including the most
  important ~~ ~~ functions ~~
}
\author{
  Paul Johnson <pauljohn@ku.edu>

  Maintainer: Paul Johnson <pauljohn@ku.edu>
}
\references{
  http://pj.freefaculty.org/R
}
\keyword{hplot}
\keyword{regression}

