% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSlopes.R
\name{se.bars}
\alias{se.bars}
\title{Draw standard error bar for discrete variables}
\usage{
se.bars(x, y, lwr, upr, width = 0.2, col, opacity = 120, lwd = 1, lty = 1)
}
\arguments{
\item{x}{The x center point}

\item{y}{The fitted "predicted" value}

\item{lwr}{The lower confidence interval bound}

\item{upr}{The upper confidence interval bound}

\item{width}{Thickness of shaded column}

\item{col}{Color for a bar}

\item{opacity}{Value in c(0, 254). 120 is default, that's partial see through.}

\item{lwd}{line width, usually 1}

\item{lty}{line type, usually 1}
}
\description{
Used with plotSlopes if plotx is discrete. This is not currently
exported.
}
\author{
Paul Johnson
}
