% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cutByTable}
\alias{cutByTable}
\title{Select most frequently occurring values from numeric or categorical variables.}
\usage{
cutByTable(x, n = 5, pct = TRUE)
}
\arguments{
\item{x}{A numeric or character variable}

\item{n}{The maximum number of values that may be returned.}

\item{pct}{Default = TRUE. Include percentage of responses within each category}
}
\value{
A named vector.
}
\description{
The "n" most frequently occurring values are returned, sorted by
frequency of occurrence (in descending order). The names attribute
includes information about the percentage of cases that have the
indicated values.
}
\details{
This is used by plotSlopes, plotCurves, and other "newdata" making
functions.
}
\author{
Paul E. Johnson <pauljohn@ku.edu>
}
