% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summary.factor}
\alias{summary.factor}
\title{Tabulates observed values and calculates entropy}
\usage{
\method{summary}{factor}(y)
}
\arguments{
\item{y}{a factor (non-numeric variable)}
}
\value{
A list, including the summary table and vector of summary
    stats if requested.
}
\description{
This adapts code from R base summary.factor. It adds
the calculation of entropy as a measure of diversity.
}
\author{
Paul E. Johnson \email{pauljohn@ku.edu}
}
