% Generated by roxygen2 (4.0.1): do not edit by hand
\name{summary.factor}
\alias{summary.factor}
\title{Tabulates observed values and calculates entropy}
\usage{
summary.factor(y, maxLevels = 5, sumstat = TRUE)
}
\arguments{
\item{y}{a factor (non-numeric variable)}

\item{maxLevels}{The maximum number of levels that will
be presented in the tabulation.}

\item{sumstat}{If TRUE (default), entropy (diversity) estimate and
the number of NAs will be returned.}
}
\value{
a vector of named elements including the summary
table as well as entropy and normed entropy.
}
\description{
This adapts code from R base summary.factor. It adds
the calculation of entropy as a measure of diversity.
}
\author{
Paul E. Johnson <pauljohn@ku.edu>
}

